refactor(cli): act on the review of the get command
Four points from review, no behaviour change beyond the argument-count message.

enum.go declared EnumSlice below the two helpers it calls. Each type now owns
one unbroken block - Enum then EnumSlice - with oneOf and allowedStrings moved
below both, so nothing is declared after a function that uses it.

get.go kept its two injection seams at the foot of the file; they move up under
the consts, where the layout convention puts vars.

printWarnings and the warning loop in definitions.go were the same code twice.
Both now go through cli/cmd/warnings.go, which also takes definitions.Warning
directly, so the map-to-string step disappears from the caller as well.

parseArgs hand-rolled the argument-count checks cobra provides; get now uses
cobra.RangeArgs(1, 2), matching definitions, which already uses MaximumNArgs.
The zero-argument message changes from "a TYPE argument is required" to cobra's
"accepts between 1 and 2 arg(s), received 0"; no test asserted either.
